backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.10.0
anthropic==0.8.1
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
chromadb==0.4.22
sentence-transformers==2.3.1
PyPDF2==3.0.1
python-docx==1.1.0
redis==5.0.1
celery==5.3.6
httpx==0.26.0
aiofiles==23.2.1
Pillow==10.2.0
